+++ src/router/history-router.ts
@@ -67,5 +67,5 @@ export class HistoryRouter implements Router {
  * Format a path with optional query parameters for use in `location.href` or the history API.
  */
 export function formatPath(path: string, query?: QueryInit): string {
-	return `${path || "/"}${query === undefined ? "" : new URLSearchParams(query)}`;
+	return `${path || "/"}${query === undefined ? "" : `?${new URLSearchParams(query)}`}`;
 }